streamlit_test/pyproject.toml
[tool.poetry]
name = "streamlit-test"
version = "0.1.0"
description = ""
authors = ["MaddoSaru <riquelme.madmonkey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.27.1"
pandas = "2.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
